Download updated firmware only

+++ update-firmware.sh
@@ -7,8 +7,16 @@ git show main:tools/info.json | jq -r '.firmwares | map_values(.pci_device_id + 
 		bin=${line%%::*}
 		url=${line#*::}
 		# echo $bin $url
-		sed -i "s|$bin::.*$|$line|g" ../PKGBUILD
-		rm ../$bin
+		old_url=$(grep "$bin::" ../PKGBUILD | awk -F:: '{print $2}')
+
+		if [ "$old_url" == "$url" ]; then
+			echo "No update for $bin"
+			continue
+		else 
+			echo "Update $bin"
+			sed -i "s|$bin::.*$|$line|g" ../PKGBUILD
+			rm ../$bin
+		fi
 	done
 popd
 updpkgsums
